requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-multipart==0.0.6
pydantic==2.4.2
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.2
pytz==2023.3